requirements.txt
amqp==1.4.9
anyjson
billiard
celery
kombu
dj-database-url
Django>=1.9,<1.10
django-celery
django-filter
django-oauth-toolkit
djangorestframework
eemeter==0.4.6
pandas>=0.18
psycopg2
redis
requests
sqlalchemy
django-grappelli
gunicorn
django-rest-swagger==0.3.5
djangorestframework-bulk
whitenoise
